requirements.txt
accelerate
datasets
httpx[http2]
mcp[cli]
pillow
pylatexenc
pyyaml
ray[default]
skypilot[aws]
torch
transformers
wandb
